+++ tests/Fixtures/i18n/fb.es.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+	'messages' => [
+		'a' => 'A-es',
+		'null-key' => null,
+	],
+];